Add options to skip archiving + remove previous archive dir

+++ actions/lib/create_archive_dir.py
@@ -7,7 +7,7 @@ def verify_src(srcdir, threshold = None):
     print("Runfolder {} is not a directory and/or doesn't exist. Aborting.".format(srcdir))
     sys.exit(1)
 
-  unaligned_link = os.path.join(os.path.sep, srcdir, "Unaligned")
+  unaligned_link = os.path.join(srcdir, "Unaligned")
   unaligned_dir = os.path.abspath(unaligned_link)
 
   if not os.path.exists(unaligned_link) or not os.path.islink(unaligned_link): 
@@ -22,17 +22,20 @@ def verify_src(srcdir, threshold = None):
     for root, dirs, files in os.walk(unaligned_dir):
       for file in files: 
         if file.lower().endswith(".fastq.gz"):
-          print(os.path.join(root, file))
           counter = counter + 1
   
     if threshold and counter < threshold: 
-        print("We only found {} files that seems to be FASTQ files. Expecting at least {} files. Looks suspicious; aborting.".format(counter, threshold))
+        print("We found only {} files that seems to be FASTQ files. Expecting at least {} files. Looks suspicious; aborting.".format(counter, threshold))
         sys.exit(1)
 
-def verify_dest(destdir): 
+def verify_dest(destdir, remove=False): 
   if os.path.exists(destdir):
-    print("Archive directory {} already exists. Will remove it first.".format(destdir))
-    shutil.rmtree(destdir)
+    if remove: 
+      print("Archive directory {} already exists. Operator requested to remove it.".format(destdir))
+      shutil.rmtree(destdir)
+    else: 
+      print("Archive directory {} already exists. Aborting.".format(destdir))
+      sys.exit(1)
 
 def create_dest(srcdir, destdir, exclude = None): 
   os.makedirs(destdir)
@@ -42,6 +45,9 @@ def create_dest(srcdir, destdir, exclude = None):
       os.symlink(os.path.join(srcdir, entry), os.path.join(destdir, entry))
  
   print("Archive directory {} created successfully.".format(destdir))
+
+def str2bool(val): 
+    return val.lower() in ("yes", "true", "1")  
 
 if __name__ == "__main__": 
   parser = argparse.ArgumentParser(description="Takes a runfolder's path as an argument and creates a copy of it with symlinks\n"\
@@ -56,12 +62,15 @@ if __name__ == "__main__":
   parser.add_argument("runfolder", help="path to the runfolder to archive")
   parser.add_argument("-e", "--exclude", action='append', help="filename to exclude from archive (argument can be repeated)")
   parser.add_argument("-t", "--threshold", default=10, type=int, help="will abort if less than this many FASTQ files are found (default 10)") 
+  parser.add_argument("-r", "--remove", default='false', type=str2bool, help="if set to true then script will remove any already existing archive directories (default false)") 
+  
   args = parser.parse_args()
   srcdir = os.path.abspath(args.runfolder)
   exclude = args.exclude
   threshold = args.threshold
+  remove = args.remove
 
   destdir = srcdir + "_archive" 
   verify_src(srcdir, threshold)
-  verify_dest(destdir)
+  verify_dest(destdir, remove)
   create_dest(srcdir, destdir, exclude)
